Попытка сделать метод Гаусса на котлине

+++ src/pro/prieran/misis/mm/two_dimension/interfaces/Func2.java
@@ -1,5 +1,5 @@
 package pro.prieran.misis.mm.two_dimension.interfaces;
 
 public interface Func2 {
-    double get(int first, int second);
+    double get(double first, double second);
 }
